docker/trainer/scripts/requirements.txt
torch==2.4.0
torchvision==0.19.0
diffusers==0.30.0
transformers==4.44.0
accelerate==0.33.0
peft==0.12.0
safetensors==0.4.4
datasets==2.20.0
Pillow==10.4.0
boto3==1.34.0
requests==2.32.0
tqdm==4.66.0
wandb==0.17.0
numpy==1.26.4
